Refactor loopback-datasource-juggler declarations
