+++ packages/juggler/lib/datasource.ts
@@ -1,6 +1,10 @@
 import {Connector} from './connector';
 
+/**
+ * DataSource denotes a configured connector
+ */
 export interface DataSource {
-  name: string,
-  connector: Connector,
+  name: string, // Name of the data source
+  connector: Connector, // The underlying connector
+  [property: string]: any; // Other properties that vary by connectors
 }
